feat: e2e encryption seems to work.. one way.
