+++ .gitignore
@@ -3,7 +3,7 @@
 # dependencies
 front-end/node_modules
 back-end/node_modules
-telegram-bot/node_modules
+telegram-bot
 front-end/package-lock.json
 back-end/package-lock.json
 /.pnp
