+++ front-end/src/components/Chat/MessageContainer.js
@@ -38,7 +38,7 @@ export default function MessageContainer({ isSender, message, fileUrl, timestamp
         }
 
         // DEV: pls comment this out else credits go byebye
-        verifyLocalisation();
+        // verifyLocalisation();
     }, []);
 
     const isAppointmentMessage = message && message.includes('appointment');
