+++ front-end/src/pages/Customer/CustomerChat.js
@@ -94,6 +94,15 @@ export default function CustomerChat() {
         // Generate the RSA Key Pair for the Customer
         RSAHandler.generateRSAKeyPair();
 
+        // Emit the User's RSA Public Key
+        RSAHandler.retrieveRSAKeyPair("rsa-public").then((res) => {
+            console.log("Sending RSA Public Key to Server");
+            socket.emit("utils:share-keys", {
+                key: res.key,
+                case: caseID
+            });
+        });
+
         // Socket.IO Event Handlers
         const handleConnection = () => {
             setIsConnected(true);
@@ -103,14 +112,6 @@ export default function CustomerChat() {
                 navigateHome();
             }
 
-            // Emit the User's RSA Public Key
-            RSAHandler.retrieveRSAKeyPair("rsa-public").then((res) => {
-                socket.emit("utils:share-keys", {
-                    key: res.key,
-                    case: caseID
-                });
-            });
-            
             socket.emit("utils:verify-activechat", customerSessionIdentifier, (chatExistanceReq) => {
                 setStaffName(chatExistanceReq.staffName);
                 if (chatExistanceReq.exist && chatExistanceReq.caseID === caseID) {
@@ -128,6 +129,7 @@ export default function CustomerChat() {
         }
 
         const handleReceiveMessage = async (msg) => {
+            // TO FIX: this will obviously fail if you are receiving the message that was sent from here (different private keys.. )
             const decryptedAESKey = await RSAHandler.decryptDataWithRSAPrivate(msg.key);
             const decryptedMessage = await AESHandler.decryptDataWithAESKey(decryptedAESKey, msg.iv, msg.fileUrl ? msg.fileUrl : msg.message);
             if (msg.fileUrl) {
@@ -171,9 +173,9 @@ export default function CustomerChat() {
 
         // Encrypt the message with the Customer's RSA Public Key
         const aesKey = await AESHandler.generateAESKey();
+        console.log(aesKey);
         const encryptedMessage = await AESHandler.encryptDataWithAESKey(isFile ? fileUrl : sentMessage, aesKey);
         const encryptedKey = await RSAHandler.encryptDataWithRSAPublic(aesKey, receiverRSAPublicKey);
-
         const formattedMsg = {
             case: caseID,
             message: isFile ? null : encryptedMessage.data,
